docs
User must generate an OpenAPI key first before their extension is able to work within the Chrome browser. This should be properly documented in the README file as this is a pain point when it comes to setting up the project. I’ve also taken out the section that states we don’t have to pay as OpenAPI keys come with billing based on rates.